aidemy-bootstrap/courses/requirements.txt
google-genai==1.0.0
soundfile==0.13.1
google-cloud-storage==2.19.0
google-cloud-pubsub==2.28.0
functions-framework==3.8.2
